sql-data/requirements.txt
dash
dash-bootstrap-components
pandas
plotly
databricks-sql-connector
databricks-connect
databricks-sdk
python-dotenv
dash-ag-grid
